requirements.txt
customtkinter
Pillow
matplotlib
pandas
sqlite3
openpyxl
ctkdateentry
ctkmessagebox
pyyaml
fpdf